+++ src/modelNames.js
@@ -1,9 +1,11 @@
+const _ = require("lodash")
+
+const isValidName = name => /^([A-Za-z$_]{1,})$/g.test(name)
 
 const checkAndRenameModelName = name => {
 
-  if (!/^([A-Za-z$_]{1,})$/g.test(name)) {
+  if (!isValidName(name)) {
     // specific replaces for TSOA 3.x
-
     if (name.includes('.'))
       name = name
         .replace(/Exclude_keyof[A-Za-z]{1,}/g, match => 'ExcludeKeys')
@@ -11,11 +13,15 @@ const checkAndRenameModelName = name => {
         .replace(/%22\~OR\~%22/g, 'Or')
         .replace(/(\.?%22)|\./g, '_')
         .replace(/__+$/, '');
+
+    if (name.includes('-'))
+      name = _.startCase(name).replace(/ /g, '')
   }
 
   return name;
 }
 
 module.exports = {
-  checkAndRenameModelName
+  checkAndRenameModelName,
+  isValidName,
 }