+++ src/routes.js
@@ -1,5 +1,6 @@
 const _ = require("lodash");
-const { parseSchema } = require("./schema");
+const { collect } = require("./utils");
+const { parseSchema, getRefType } = require("./schema");
 const { checkAndRenameModelName } = require("./modelNames");
 const { inlineExtraFormatters } = require("./typeFormatters");
 
@@ -18,9 +19,12 @@ const getTypeFromRequestInfo = (requestInfo, parsedSchemas, contentType) => {
   if (schema) {
     const extractedSchema = _.get(schema, 'additionalProperties', schema);
     const { content } = parseSchema(extractedSchema, 'none', inlineExtraFormatters);
-    const foundSchema = _.find(parsedSchemas, parsedSchema => _.isEqual(parsedSchema.content, content))
+    const foundedSchemaByName = _.find(parsedSchemas, parsedSchema => parsedSchema.name === content)
+    const foundSchemaByContent = _.find(parsedSchemas, parsedSchema => _.isEqual(parsedSchema.content, content))
 
-    return foundSchema ? foundSchema.name : checkAndRenameModelName(content);
+    const foundSchema = foundedSchemaByName || foundSchemaByContent
+
+    return checkAndRenameModelName(foundSchema ? foundSchema.name : content);
   }
 
   return 'any';
@@ -43,9 +47,10 @@ const getRouteName = (operationId, method, route, moduleName) => {
   return createCustomOperationId(method, route, moduleName);
 }
 
-const parseRoutes = (routes, parsedSchemas) =>
+const parseRoutes = (routes, parsedSchemas, components) =>
   _.entries(routes)
     .reduce((routes, [route, requestInfoByMethodsMap]) => {
+      const globalParametersMap = _.get(components, "parameters", {});
       parameters = _.get(requestInfoByMethodsMap, 'parameters');
 
       // TODO: refactor that hell
@@ -74,13 +79,26 @@ const parseRoutes = (routes, parsedSchemas) =>
             responses,
           } = requestInfo;
           const hasSecurity = !!(security && security.length);
-          const pathParams = _.filter(parameters, parameter => parameter.in === 'path');
-          const queryParams = _.filter(parameters, parameter => parameter.in === 'query');
+          const pathParams = collect(parameters, parameter => {
+            if (parameter.in === 'path') return parameter;
+            
+            const refTypeName = getRefType(parameter);
+            const globalParam = refTypeName && globalParametersMap[refTypeName]
+            return globalParam && globalParametersMap[refTypeName].in === "path" && globalParam
+          })
+          const queryParams = collect(parameters, parameter => {
+            if (parameter.in === 'query') return parameter;
+            
+            const refTypeName = getRefType(parameter);
+            const globalParam = refTypeName && globalParametersMap[refTypeName]
+            return globalParam && globalParametersMap[refTypeName].in === "query" && globalParam;
+          })
           const moduleName = _.camelCase(route.split('/').filter(Boolean)[0]);
 
           const routeName = getRouteName(operationId, method, route, moduleName);
-          
-          const queryObjectSchema = queryParams.length && queryParams.reduce((objectSchema, queryPartSchema) => ({
+          const name = _.camelCase(routeName);
+
+          const queryObjectSchema = _.reduce(queryParams, (objectSchema, queryPartSchema) => ({
             ...objectSchema,
             properties: {
               ...objectSchema.properties,
@@ -93,18 +111,26 @@ const parseRoutes = (routes, parsedSchemas) =>
 
           const bodyParamName = requestInfo.requestBodyName || (requestBody && requestBody.name) || "data";
 
+          const queryType = queryParams.length
+            ? parseSchema(queryObjectSchema, null, inlineExtraFormatters).content
+            : null;
+
+          const bodyType = requestBody
+            ? getTypeFromRequestInfo(requestBody, parsedSchemas, "application/json")
+            : null;
+
           const args = [
             ...(pathParams.map(param => ({
               name: param.name,
               type: parseSchema(param.schema, null, inlineExtraFormatters).content
             }))),
-            queryParams.length && {
+            queryType && {
               name: 'query',
-              type: parseSchema(queryObjectSchema, null, inlineExtraFormatters).content,
+              type: queryType,
             },
-            requestBody && {
+            bodyType && {
               name: bodyParamName,
-              type: getTypeFromRequestInfo(requestBody, parsedSchemas, "application/json"),
+              type: bodyType,
             },
           ].filter(Boolean)
 
@@ -139,10 +165,13 @@ const parseRoutes = (routes, parsedSchemas) =>
           ].filter(Boolean);
 
           return {
-            moduleName,
+            moduleName: _.replace(moduleName, /^(\d)/, 'v$1'),
             security: hasSecurity,
             hasQuery: !!queryParams.length,
-            name: _.camelCase(routeName),
+            queryType: queryType || '{}',
+            bodyType: bodyType || 'never',
+            name,
+            pascalName: _.upperFirst(name),
             comments,
             args,
             method: _.upperCase(method),
@@ -173,9 +202,11 @@ const groupRoutes = routes => {
         )
         route.comments.push(`@originalName ${route.name}`)
         route.comments.push(`@duplicate`)
-        route.name += ++duplicates[route.moduleName][route.name];
+        const duplicateNumber = ++duplicates[route.moduleName][route.name]
+        route.name += duplicateNumber;
+        route.pascalName += duplicateNumber;
       }
-      
+
       modules[route.moduleName].push(route)
     } else {
       modules.$outOfModule.push(route)