+++ tests/helpers/createGeneratedApiInfos.js
@@ -1,0 +1,6 @@
+const fs = require("fs");
+const { resolve } = require("path");
+
+module.exports = (pathToApis) =>
+  (fs.readdirSync(pathToApis) || [])
+    .map(fileName => resolve(pathToApis, fileName))