+++ tests/helpers/specGenerateOptions.js
@@ -1,0 +1,6 @@
+const path = require("path")
+
+module.exports = {
+  name: "Api.ts",
+  input: path.resolve(__dirname, "../schemas/v2.0/petstore-simple.yaml"),
+}