+++ tests/validate.js
@@ -1,42 +1,25 @@
-const fs = require("fs");
-const path = require("path");
-const tsc = require("typescript");
+const { resolve } = require("path");
+const createGeneratedApiInfos = require("./helpers/createGeneratedApiInfos");
+const validateGeneratedModule = require("./helpers/validateGeneratedModule");
 
-const createGeneratedApiInfos = (pathToApis) =>
-  (fs.readdirSync(pathToApis) || [])
-    .map(fileName => path.resolve(pathToApis, fileName))
+const v2ApiPaths = createGeneratedApiInfos(resolve(__dirname, "./generated/v2.0"));
+const v3ApiPaths = createGeneratedApiInfos(resolve(__dirname, "./generated/v3.0"));
 
-const v2ApiPaths = createGeneratedApiInfos(path.resolve(__dirname, "./generated/v2.0"));
-const v3ApiPaths = createGeneratedApiInfos(path.resolve(__dirname, "./generated/v3.0"));
+let hasErrors = false;
 
 [
   ...v2ApiPaths,
   ...v3ApiPaths,
-].forEach(path => {
-  process.stdout.write(`validating ${path}: `)
-  var program = tsc.createProgram([path], {
-    noEmitOnError: true,
-    noImplicitAny: false,
-    target: tsc.ScriptTarget.ES5, module: tsc.ModuleKind.CommonJS
-  });
-  var emitResult = program.emit(void 0, void 0, void 0, true,)
+].forEach(pathToFile => {
+  const diagnostics = validateGeneratedModule({ pathToFile })
 
-  var allDiagnostics = emitResult.diagnostics;
-
-  allDiagnostics.forEach(diagnostic => {
-      var message = tsc.flattenDiagnosticMessageText(diagnostic.messageText, '\n');
-      if (!diagnostic.file) {
-          console.log(message);
-          return;
-      }
-      var { line, character } = diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start);
-      console.log(`${diagnostic.file.fileName} (${line + 1},${character + 1}): ${message}`);
-  });
-
-  process.stdout.write(`errors ${emitResult.diagnostics.length}\r\n`)
-  if (emitResult.diagnostics.length){
-    process.exit(1)
+  if (!hasErrors) {
+    hasErrors = !!diagnostics.length
   }
 })
 
-console.log('everything is good:)')
+if (hasErrors) {
+  process.exit(1)
+} else {
+  console.log('everything is good:)')
+}